// frontend/src/api/generated/model/maintenanceCategory.ts
export type MaintenanceCategory = typeof MaintenanceCategory[keyof typeof MaintenanceCategory];


export const MaintenanceCategory = {
  plumbing: 'plumbing',
  electrical: 'electrical',
  furniture: 'furniture',
  heating: 'heating',
  network: 'network',
  other: 'other',
} as const;