// frontend/src/api/generated/model/maintenanceLocation.ts
export type MaintenanceLocation = typeof MaintenanceLocation[keyof typeof MaintenanceLocation];


export const MaintenanceLocation = {
  own_room: 'own_room',
  common_area: 'common_area',
} as const;